+++ lib/sidekiq/fetch.rb
@@ -13,9 +13,8 @@ module Sidekiq
     TIMEOUT = 1
 
     def initialize(mgr, options)
-      klass = Sidekiq.options[:fetch] || BasicFetch
       @mgr = mgr
-      @strategy = klass.new(options)
+      @strategy = Fetcher.strategy.new(options)
     end
 
     # Fetching is straightforward: the Manager makes a fetch
@@ -57,6 +56,10 @@ module Sidekiq
     def self.done?
       @done
     end
+
+    def self.strategy
+      Sidekiq.options[:fetch] || BasicFetch
+    end
   end
 
   class BasicFetch
@@ -75,13 +78,13 @@ module Sidekiq
       Sidekiq.logger.debug { "Re-queueing terminated jobs" }
       jobs_to_requeue = {}
       inprogress.each do |unit_of_work|
-        jobs_to_requeue[unit_of_work.queue] ||= []
-        jobs_to_requeue[unit_of_work.queue] << unit_of_work.message
+        jobs_to_requeue[unit_of_work.queue_name] ||= []
+        jobs_to_requeue[unit_of_work.queue_name] << unit_of_work.message
       end
 
       Sidekiq.redis do |conn|
         jobs_to_requeue.each do |queue, jobs|
-          conn.rpush(queue, jobs)
+          conn.rpush("queue:#{queue}", jobs)
         end
       end
       Sidekiq.logger.info("Pushed #{inprogress.size} messages back to Redis")
@@ -98,7 +101,7 @@ module Sidekiq
 
       def requeue
         Sidekiq.redis do |conn|
-          conn.rpush(queue, message)
+          conn.rpush("queue:#{queue_name}", message)
         end
       end
     end