+++ test/helper.rb
@@ -1,7 +1,14 @@
+require 'coveralls'
+Coveralls.wear! do
+  add_filter "/test/"
+end
+
 ENV['RACK_ENV'] = ENV['RAILS_ENV'] = 'test'
 if ENV.has_key?("SIMPLECOV")
   require 'simplecov'
-  SimpleCov.start
+  SimpleCov.start do
+    add_filter "/test/"
+  end
 end
 
 begin
@@ -10,8 +17,8 @@ rescue LoadError
 end
 
 require 'minitest/unit'
-require 'minitest/pride'
 require 'minitest/autorun'
+require 'minitest/emoji'
 
 require 'sidekiq'
 require 'sidekiq/util'